requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
importlib-metadata==1.7.0
kombu==4.6.11
pytz==2020.1
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
